Very large change
rename app.py to CHIME.py so the title of the page is correct
update docs with new main file name
add chime directory with models, presentation, and utils - for general cleaning and in preparation of making this a module. This changes pretty much everything, but I think it makes future changes easier
