+++ docs/contributing/app-dev.md
@@ -24,7 +24,7 @@ See [Streamlit's Getting Started guide](https://docs.streamlit.io/getting_starte
 ```bash
 pipenv shell
 pipenv install
-streamlit run app.py
+streamlit run CHIME.py
 ```
 
 ### With `conda`
@@ -33,7 +33,7 @@ streamlit run app.py
 conda env create -f environment.yml
 source activate chime
 pip install streamlit
-streamlit run app.py
+streamlit run CHIME.py
 ```
 
 ### Choosing a Different Port
@@ -42,7 +42,7 @@ If you need to run the application on a different port than the default (8000), 
 
 ```bash
 export STREAMLIT_SERVER_PORT=1234
-streamlit run app.py
+streamlit run CHIME.py
 ```
 
 ## Project Layout
@@ -50,7 +50,7 @@ streamlit run app.py
 ### Application files
 
 - `script/`: Developer workflow scripts following [GitHub's Scripts To Rule Them All](https://github.com/github/scripts-to-rule-them-all) pattern.
-- `app.py`: Main source for the application
+- `CHIME.py`: Main source for the application
 - `.streamlit/`: [Streamlit config options](https://docs.streamlit.io/cli.html)
 - `.env`: Local environment variables to use when running application, this file is copied from `.env.example` to start you out and then ignored by git
 - `Pipfile`
